requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.7

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-yasg==1.21.7

# Filtering
django-filter==23.3

# AI/ML
scikit-learn==1.3.1
pandas==2.1.4
numpy==1.24.4
joblib==1.3.2

# Reports
reportlab==4.0.6
openpyxl==3.1.2

# Payments
stripe==7.7.0

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2